requirements.txt
autopep8
pylint
docopt
pyyaml
pyglet
